moved mac addr into the wotail lib
